fix bug, restore memory.cpp to old version

+++ lightseq/csrc/pybind/pybind_model.cpp
@@ -196,7 +196,7 @@ class PyBertCrf {
     }
   }
 
-  py::array_t<float> infer(
+  py::array_t<int> infer(
       py::array_t<int, py::array::c_style | py::array::forcecast> input_seq) {
     auto input_seq_out = input_seq.mutable_unchecked<2>();
     const int *input_seq_data = input_seq_out.data(0, 0);
